fix: 실행 오류 해결 (빈 중복된 것 해결), refactor: LikeService 메서드 수정

+++ src/main/java/com/example/recipeapp/domain/like/controller/dto/LikeCountResponseDto.java
@@ -9,5 +9,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class LikeCountResponseDto {
     private Long recipeId;
-    private Long likesCount;
+    private int likesCount;
 }
